refactor: renamed func
