+++ packages/deployment-service/src/utils/index.ts
@@ -1,4 +1,4 @@
 export * from './ownership'
 export * from './resolve-customer'
-export * from './resolve-developer-id'
+export * from './resolve-creds'
 export * from './recurse-dir'
